+++ src/utils/data/vatglasses.ts
@@ -38,7 +38,7 @@ export interface VatglassesActiveData {
 }
 
 export interface VatglassesActivePosition {
-    atc: VatsimShortenedController;
+    atc: VatsimShortenedController[];
     sectors: TurfFeature<TurfPolygon>[] | null;
     sectorsCombined: TurfFeature<TurfPolygon>[] | null;
     airspaceKeys: string | null;
@@ -67,7 +67,7 @@ export interface VatglassesSectorProperties {
     max: number;
     countryGroupId: string;
     vatglassesPositionId: string;
-    atc: VatsimShortenedController;
+    atc: VatsimShortenedController[];
     colour: string;
     type: 'vatglasses';
 }
@@ -101,7 +101,7 @@ function updateVatglassesPositionsAndAirspaces() {
         };
     }
 
-    const vatglassesActiveController: { [countryGroupId: string]: { [vatglassesPositionId: string]: VatsimShortenedController } } = {}; // countryGroupId is the name of the json files which are split into areas
+    const vatglassesActiveControllers: { [countryGroupId: string]: { [vatglassesPositionId: string]: VatsimShortenedController[] } } = {}; // countryGroupId is the name of the json files which are split into areas
     const fallbackPositions: VatsimShortenedController[] = [];
     // Fill the vatglassesActiveStations object with the active stations
     if (vatglassesData) {
@@ -148,7 +148,6 @@ function updateVatglassesPositionsAndAirspaces() {
             }
 
             let foundMatchingVatglassesController = false;
-            let doublePositionMatch = false;
             // We sort the keys with the first two chars of the callsign, so we can check first the most likely countryGroupIds
             const keys = Object.keys(vatglassesData);
             const searchString = atc.callsign.substring(0, 2);
@@ -164,21 +163,20 @@ function updateVatglassesPositionsAndAirspaces() {
                     if (vatglassesPosition.frequency && vatglassesPosition.frequency !== atc.frequency) continue;
                     if (!atc.callsign.endsWith(vatglassesPosition.type)) continue;
                     if (!vatglassesPosition.pre.some((prefix: string) => atc.callsign.startsWith(prefix))) continue;
-                    if (vatglassesActiveController?.[countryGroupId]?.[vatglassesPositionId] != null) {
-                        doublePositionMatch = true;
-                        continue; // There is already a controller assigned to this vatglasses position
+
+                    if (!vatglassesActiveControllers[countryGroupId]) {
+                        vatglassesActiveControllers[countryGroupId] = {};
                     }
 
-                    if (!vatglassesActiveController[countryGroupId]) {
-                        vatglassesActiveController[countryGroupId] = {};
-                    }
-                    vatglassesActiveController[countryGroupId][vatglassesPositionId] = atc;
+                    vatglassesActiveControllers[countryGroupId][vatglassesPositionId] ??= [];
+
+                    vatglassesActiveControllers[countryGroupId][vatglassesPositionId].push(atc);
                     foundMatchingVatglassesController = true;
                     break;
                 }
             }
 
-            if (!foundMatchingVatglassesController && !doublePositionMatch) {
+            if (!foundMatchingVatglassesController) {
                 fallbackPositions.push(atc);
             }
         }
@@ -192,7 +190,7 @@ function updateVatglassesPositionsAndAirspaces() {
         const countryGroup = vatglassesData[countryGroupId];
 
         let activeGroupVatglassesPosition: string[] = [];
-        if (vatglassesActiveController[countryGroupId]) activeGroupVatglassesPosition = Object.keys(vatglassesActiveController[countryGroupId]);
+        if (vatglassesActiveControllers[countryGroupId]) activeGroupVatglassesPosition = Object.keys(vatglassesActiveControllers[countryGroupId]);
         for (const airspaceIndex in countryGroup.airspace) {
             const airspace = countryGroup.airspace[airspaceIndex];
             let airspaceOwner: string [] = [];
@@ -208,8 +206,8 @@ function updateVatglassesPositionsAndAirspaces() {
                     const elementSplit = element.split('/');
                     const otherCountryCode = elementSplit[0];
                     const otherGroupVatglassesPosition = elementSplit[1];
-                    if (vatglassesActiveController[otherCountryCode]) {
-                        const otherGroupVatglassesActivePositions = Object.keys(vatglassesActiveController[otherCountryCode]);
+                    if (vatglassesActiveControllers[otherCountryCode]) {
+                        const otherGroupVatglassesActivePositions = Object.keys(vatglassesActiveControllers[otherCountryCode]);
                         return otherGroupVatglassesActivePositions.includes(otherGroupVatglassesPosition);
                     }
                 }
@@ -238,12 +236,13 @@ function updateVatglassesPositionsAndAirspaces() {
     }
 
     // It is possible that a position is defined in VatGlasses data but no airspaces are defined for it. In this case, it is not added to the fallbackPositions above, because the position exists, but no airspace is defined for it. We have to add it to the fallbackPositions here.
-    for (const countryGroupId in vatglassesActiveController) {
-        for (const positionId in vatglassesActiveController[countryGroupId]) {
+    for (const countryGroupId in vatglassesActiveControllers) {
+        for (const positionId in vatglassesActiveControllers[countryGroupId]) {
             if (!vatglassesActiveAirspaces[countryGroupId] || !vatglassesActiveAirspaces[countryGroupId][positionId] || Object.keys(vatglassesActiveAirspaces[countryGroupId][positionId]).length === 0) {
-                const controller = vatglassesActiveController[countryGroupId][positionId];
-                if (!fallbackPositions.some(fallback => fallback.callsign === controller.callsign)) {
-                    fallbackPositions.push(controller);
+                const controllers = vatglassesActiveControllers[countryGroupId][positionId];
+                const fallback = controllers.filter(x => fallbackPositions.some(fallback => fallback.callsign === x.callsign));
+                if (!fallback.length) {
+                    fallbackPositions.push(...controllers);
                 }
             }
         }
@@ -290,10 +289,10 @@ function updateVatglassesPositionsAndAirspaces() {
             }
             else {
                 if (mode === 'server') {
-                    newVatglassesActivePositions[countryGroupId][positionId] = { atc: vatglassesActiveController[countryGroupId][positionId], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: null }; // we set null instead of [], because null is the signal for later that we have to recalculate it
+                    newVatglassesActivePositions[countryGroupId][positionId] = { atc: vatglassesActiveControllers[countryGroupId][positionId], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: null }; // we set null instead of [], because null is the signal for later that we have to recalculate it
                 }
                 else {
-                    newVatglassesActivePositions[countryGroupId][positionId] = { atc: vatglassesActiveController[countryGroupId][positionId], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: ref<string | null>(null) }; // we set null instead of [], because null is the signal for later that we have to recalculate it
+                    newVatglassesActivePositions[countryGroupId][positionId] = { atc: vatglassesActiveControllers[countryGroupId][positionId], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: ref<string | null>(null) }; // we set null instead of [], because null is the signal for later that we have to recalculate it
                 }
             }
 
@@ -322,10 +321,10 @@ function updateVatglassesPositionsAndAirspaces() {
     newVatglassesActivePositions['fallback'] = {};
     for (const atc of fallbackPositions) {
         if (mode === 'server') {
-            newVatglassesActivePositions['fallback'][atc.callsign] = { atc: atc, sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: null };
+            newVatglassesActivePositions['fallback'][atc.callsign] = { atc: [atc], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: null };
         }
         else {
-            newVatglassesActivePositions['fallback'][atc.callsign] = { atc: atc, sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: ref<string | null>(null) };
+            newVatglassesActivePositions['fallback'][atc.callsign] = { atc: [atc], sectors: null, sectorsCombined: null, airspaceKeys: null, lastUpdated: ref<string | null>(null) };
         }
     }
 
@@ -393,7 +392,7 @@ function getActiveSectorsOfAirspace(airspace: VatglassesAirspace) {
 
 
 // Converts from vatglasses sector format to geojson format
-function convertSectorToGeoJson(sector: VatglassesSector, countryGroupId: string, positionId: string, atc: VatsimShortenedController, positions: VatglassesActivePositions) {
+function convertSectorToGeoJson(sector: VatglassesSector, countryGroupId: string, positionId: string, atc: VatsimShortenedController[], positions: VatglassesActivePositions) {
     const vatglassesData = dataStore?.vatglasses?.value?.data ?? radarStorage.vatglasses?.data.data;
 
     try {